refactor(axum-kbve): split proxy.rs godfile into proxy/ module tree
Break the 3229-line transport/proxy.rs into a proxy/ submodule tree with
no behavior change:

  proxy/core.rs        ServiceProxy engine (handle/forward_request/forward),
                       auth gates, extract_auth_token, reqwest_headers
  proxy/simple.rs      Grafana, ArgoCD, Forgejo, Factorio, Edge
  proxy/clickhouse.rs  ClickHouse direct route
  proxy/kubevirt.rs    KubeVirt proxy + VNC handlers
  proxy/kasm.rs        KASM (pw cache, WS bridge, launch-URL SSRF policy)
  proxy/firecracker.rs Firecracker + Firecracker-Net + account resolution
  proxy/windmill.rs    Windmill
  proxy/vibeshine.rs   Vibeshine + WebRTC relay
  proxy/guacamole.rs   Guacamole + WS bridge
  proxy/chuckrpg.rs    multi-tenant ROWS registry

Engine + gates become pub(super) in core; mod.rs glob-re-exports every
pub item so external call sites (main.rs, https.rs, openapi.rs,
forgejo_api.rs) resolve unchanged. Cross-module items (KUBEVIRT static,
KASM_NAMESPACE) exported explicitly; vnc_hub refs repathed to
crate::transport::vnc_hub. Tests split: extract_auth_token -> core,
url_passes_policy -> kasm.

Pure relocation — cargo check + clippy clean, 13 proxy tests pass.
Part of #14269.
