+++ apps/kbve/axum-kbve/src/transport/proxy/simple.rs
@@ -1,0 +1,268 @@
+use std::sync::OnceLock;
+use std::time::Duration;
+
+use axum::{
+    body::{Body},
+    extract::{Path, Request},
+    http::{StatusCode},
+    response::{IntoResponse, Response},
+};
+use reqwest::Client;
+use serde_json::json;
+use tracing::{debug, warn};
+
+use super::core::*;
+
+static GRAFANA: OnceLock<ServiceProxy> = OnceLock::new();
+
+pub fn init_grafana_proxy() -> bool {
+    let upstream = match std::env::var("GRAFANA_UPSTREAM_URL") {
+        Ok(u) => u.trim_end_matches('/').to_string(),
+        Err(_) => return false,
+    };
+
+    let client = Client::builder()
+        .redirect(reqwest::redirect::Policy::none())
+        .connect_timeout(Duration::from_secs(5))
+        .timeout(Duration::from_secs(15))
+        .build()
+        .expect("failed to build reqwest client for grafana proxy");
+
+    GRAFANA
+        .set(ServiceProxy {
+            name: "Grafana",
+            client,
+            upstream,
+            upstream_token: None,
+            upstream_headers: Vec::new(),
+            iframe_safe: false,
+            streaming: false,
+        })
+        .is_ok()
+}
+
+pub async fn grafana_proxy_handler(path: Option<Path<String>>, req: Request<Body>) -> Response {
+    match GRAFANA.get() {
+        Some(proxy) => proxy.handle(path, req).await,
+        None => (
+            StatusCode::SERVICE_UNAVAILABLE,
+            axum::Json(json!({"error": "Grafana proxy not configured"})),
+        )
+            .into_response(),
+    }
+}
+
+static ARGO: OnceLock<ServiceProxy> = OnceLock::new();
+
+pub fn init_argo_proxy() -> bool {
+    let upstream = match std::env::var("ARGOCD_UPSTREAM_URL") {
+        Ok(u) => u.trim_end_matches('/').to_string(),
+        Err(_) => return false,
+    };
+
+    let auth_token = match std::env::var("ARGOCD_AUTH_TOKEN") {
+        Ok(t) => t,
+        Err(_) => return false,
+    };
+
+    let mut builder = Client::builder()
+        .redirect(reqwest::redirect::Policy::none())
+        .connect_timeout(Duration::from_secs(10))
+        .timeout(Duration::from_secs(120));
+
+    if let Ok(ca_path) = std::env::var("ARGOCD_CA_CERT_PATH") {
+        match std::fs::read(&ca_path) {
+            Ok(pem) => match reqwest::Certificate::from_pem(&pem) {
+                Ok(cert) => {
+                    builder = builder.add_root_certificate(cert);
+                    debug!("loaded ArgoCD CA certificate from {ca_path}");
+                }
+                Err(e) => {
+                    warn!("failed to parse ArgoCD CA cert at {ca_path}: {e}");
+                    return false;
+                }
+            },
+            Err(e) => {
+                warn!("failed to read ArgoCD CA cert at {ca_path}: {e}");
+                return false;
+            }
+        }
+    }
+
+    let client = builder
+        .build()
+        .expect("failed to build reqwest client for argo proxy");
+
+    ARGO.set(ServiceProxy {
+        name: "ArgoCD",
+        client,
+        upstream,
+        upstream_token: Some(auth_token),
+        upstream_headers: Vec::new(),
+        iframe_safe: false,
+        streaming: false,
+    })
+    .is_ok()
+}
+
+pub async fn argo_proxy_handler(path: Option<Path<String>>, req: Request<Body>) -> Response {
+    match ARGO.get() {
+        // Method-aware: GET reads require DASHBOARD_VIEW; mutating verbs
+        // (sync, refresh-via-POST, delete) require DASHBOARD_MANAGE.
+        Some(proxy) => proxy.handle_method_aware(path, req).await,
+        None => (
+            StatusCode::SERVICE_UNAVAILABLE,
+            axum::Json(json!({"error": "ArgoCD proxy not configured"})),
+        )
+            .into_response(),
+    }
+}
+
+static FORGEJO: OnceLock<ServiceProxy> = OnceLock::new();
+
+pub fn init_forgejo_proxy() -> bool {
+    let upstream = match std::env::var("FORGEJO_UPSTREAM_URL") {
+        Ok(u) => u.trim_end_matches('/').to_string(),
+        Err(_) => return false,
+    };
+
+    let auth_token = match std::env::var("FORGEJO_AUTH_TOKEN") {
+        Ok(t) => t,
+        Err(_) => return false,
+    };
+
+    let client = Client::builder()
+        .redirect(reqwest::redirect::Policy::none())
+        .connect_timeout(Duration::from_secs(5))
+        .timeout(Duration::from_secs(15))
+        .build()
+        .expect("failed to build reqwest client for forgejo proxy");
+
+    FORGEJO
+        .set(ServiceProxy {
+            name: "Forgejo",
+            client,
+            upstream,
+            upstream_token: Some(auth_token),
+            upstream_headers: Vec::new(),
+            iframe_safe: false,
+            streaming: false,
+        })
+        .is_ok()
+}
+
+pub async fn forgejo_proxy_handler(path: Option<Path<String>>, req: Request<Body>) -> Response {
+    match FORGEJO.get() {
+        Some(proxy) => proxy.handle_method_aware(path, req).await,
+        None => (
+            StatusCode::SERVICE_UNAVAILABLE,
+            axum::Json(json!({"error": "Forgejo proxy not configured"})),
+        )
+            .into_response(),
+    }
+}
+
+static FACTORIO: OnceLock<ServiceProxy> = OnceLock::new();
+
+pub fn init_factorio_proxy() -> bool {
+    let upstream = std::env::var("FACTORIO_CTL_URL")
+        .unwrap_or_else(|_| "http://factorio-ctl.factorio.svc.cluster.local:9002".into());
+
+    let client = Client::builder()
+        .redirect(reqwest::redirect::Policy::none())
+        .connect_timeout(Duration::from_secs(10))
+        .timeout(Duration::from_secs(30))
+        .build()
+        .expect("failed to build reqwest client for factorio proxy");
+
+    FACTORIO
+        .set(ServiceProxy {
+            name: "Factorio",
+            client,
+            upstream: upstream.trim_end_matches('/').to_string(),
+            upstream_token: None,
+            upstream_headers: Vec::new(),
+            iframe_safe: false,
+            streaming: false,
+        })
+        .is_ok()
+}
+
+pub async fn factorio_proxy_handler(path: Option<Path<String>>, req: Request<Body>) -> Response {
+    let headers = req.headers().clone();
+    if let Err(resp) = require_dashboard_view(&headers, "Factorio").await {
+        return resp;
+    }
+
+    match FACTORIO.get() {
+        Some(proxy) => proxy.handle(path, req).await,
+        None => (
+            StatusCode::SERVICE_UNAVAILABLE,
+            axum::Json(json!({"error": "Factorio proxy not configured"})),
+        )
+            .into_response(),
+    }
+}
+
+/// Public read-only proxy for the factorio-ctl OpenAPI document. Mirrors
+/// [`firecracker_openapi_handler`]; the operations stay staff-gated through
+/// [`factorio_proxy_handler`].
+pub async fn factorio_openapi_handler(req: Request<Body>) -> Response {
+    match FACTORIO.get() {
+        Some(proxy) => {
+            proxy
+                .handle_preauthorized(Some(Path("openapi.json".to_string())), req)
+                .await
+        }
+        None => (
+            StatusCode::SERVICE_UNAVAILABLE,
+            axum::Json(json!({"error": "Factorio proxy not configured"})),
+        )
+            .into_response(),
+    }
+}
+
+static EDGE: OnceLock<ServiceProxy> = OnceLock::new();
+
+pub fn init_edge_proxy() -> bool {
+    let supabase_url = match std::env::var("SUPABASE_URL") {
+        Ok(u) => u.trim_end_matches('/').to_string(),
+        Err(_) => return false,
+    };
+    let upstream = format!("{supabase_url}/functions/v1");
+
+    let service_role_key = match std::env::var("SUPABASE_SERVICE_ROLE_KEY") {
+        Ok(k) => k,
+        Err(_) => return false,
+    };
+
+    let client = Client::builder()
+        .redirect(reqwest::redirect::Policy::none())
+        .connect_timeout(Duration::from_secs(5))
+        .timeout(Duration::from_secs(15))
+        .build()
+        .expect("failed to build reqwest client for edge proxy");
+
+    EDGE.set(ServiceProxy {
+        name: "Edge",
+        client,
+        upstream,
+        upstream_token: Some(service_role_key),
+        upstream_headers: Vec::new(),
+        iframe_safe: false,
+        streaming: false,
+    })
+    .is_ok()
+}
+
+pub async fn edge_proxy_handler(path: Option<Path<String>>, req: Request<Body>) -> Response {
+    match EDGE.get() {
+        Some(proxy) => proxy.handle(path, req).await,
+        None => (
+            StatusCode::SERVICE_UNAVAILABLE,
+            axum::Json(json!({"error": "Edge proxy not configured"})),
+        )
+            .into_response(),
+    }
+}
+
